refactor: joining login and registration modules
the two modules weren't complete on their own and used a lot of components from each other, infact the registration module didn't have any pages or any containers and used them from the login module and login module didn't have a module configuration
these were originally one module named login
