+++ web-client/src/modules/login/index.ts
@@ -1,6 +1,6 @@
 import { Module } from 'src/types/module';
 
-import ContentPage from '../login/pages/ContentPage';
+import ContentPage from './pages/ContentPage';
 
 const loginModule: Module = {
   path: '/login',
